Updating migration scripts to get clean migration on Rails 7 + MySQL 8

+++ db/migrate/20170719000343_create_diary_entries.rb
@@ -5,8 +5,8 @@ class CreateDiaryEntries < ActiveRecord::Migration[5.1]
       t.text :notes
     end
 
-    add_reference(:diary_entries, :candidate, type: :integer, foreign_key: true)
+    add_reference(:diary_entries, :candidate, type: :bigint, foreign_key: true)
     add_reference(:diary_entries, :diary_entry_type, foreign_key: true)
-    add_reference(:diary_entries, :user, type: :integer, foreign_key: true)
+    add_reference(:diary_entries, :user, type: :bigint, foreign_key: true)
   end
 end
